split out the delta phase
